All in one effort to conclude the architectural changes to YMU and conform to the recommended Sencha MVC framework in ExtJS4.
In this MVC effort, several files where moved into their respective
directories/roles within the MVC architecture. Although some are still
pending to take advantage of the controllers. Component logic for
modifying the view should be in the controllers. Other changes to this
commit include the footer dynamic loader. It also uses the latest MVC
framework, it should be use as a model.